% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf-pmmeans-class-generic.R
\docType{methods}
\name{as.shinystan,nauf.pmm.list-method}
\alias{as.shinystan,nauf.pmm.list-method}
\title{Create a shinystan object for posterior marginal means from nauf models.}
\usage{
\S4method{as.shinystan}{nauf.pmm.list}(X, ...)
}
\arguments{
\item{X}{A \code{\link{nauf.pmm.list}}.}

\item{...}{Not used.}
}
\value{
A \code{\link[shinystan]{shinystan}} object based on the posterior
marginal means (and possibly pairwise comparisons) in \code{X}.  This allows
the results to be viewed in the \code{shinystan} GUI as if they had been
included in the generated quantities block of the model.
}
\description{
Joins the \code{samples} elements from the \code{\link{nauf.pmm.stan}} objects
in a \code{\link{nauf.pmm.list}}, adds the \code{log-posterior} to them,
and passes the resulting array to the \code{array} method for
\code{\link[shinystan]{as.shinystan}}.
}
\seealso{
\code{\link[shinystan]{launch_shinystan}}
}

