% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf.stanreg-class-generic.R
\docType{methods}
\name{as.shinystan,nauf.stanreg-method}
\alias{as.shinystan,nauf.stanreg-method}
\title{Create a shinystan object from a nauf.stanreg model.}
\usage{
\S4method{as.shinystan}{nauf.stanreg}(X, ...)
}
\arguments{
\item{X}{A \code{\link{nauf.stanreg}} model.}

\item{...}{See \code{\link[shinystan]{as.shinystan}}.}
}
\value{
See \code{\link[shinystan]{as.shinystan}}.
}
\description{
This is a simple wrapper function that calls the \code{\link[shinystan]{as.shinystan}}
method for \code{stanreg} objects.
}

