% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf-package.R
\docType{package}
\name{foo}
\alias{foo}
\alias{foo-package}
\title{nauf: An R package for fitting regressions with not applicable values in unordered factors.}
\description{
The \code{nauf} package provides tools for fitting regressions where
unordered factors may have \code{NA} values in addition to their
regular levels.  In this package, \code{NA} is used to encode when an
unordered factor is truly \emph{not applicable}.  This is different
than "not available" or "missing at random".  The concept applies only to
unordered factors, and indicates that the factor is simply not meaningful
for an observation, or that while the observation may technically be
definable by one of the factor levels, the interpretation of its belonging
to that group isn't the same.
}
\details{
For example, in a linguistic study examining
three dialects of the same language, where two dialects have both monolingual
and bilingual speakers, but the third has only monolingual speakers, a binary
factor \code{bilingual} with levels \code{FALSE} and \code{TRUE} doesn't
apply to the third dialect.  While you could code all observations from
the third dialect as \code{FALSE}, the meaning would not be the same,
since in the first two dialects \code{FALSE} means
\emph{monolingual as opposed to bilingual}, while in the third it does not
have this meaning.  In this case, the factor could be coded as \code{NA}
for observations from the third dialect, and the tools in the \code{nauf}
package could be used to include them in a regression.  They are included
by using sum contrasts for unordered factors, allowing \code{NA} values
to pass to the model matrix, and then setting all \code{NA} values to zero
(in sociolinguistics this is called factor slashing).  The treatment of
unordered factors interacting with these \code{NA} values is more complex
and can often require contrast changes (see \code{\link{nauf_interaction}}).

The \code{nauf} package works by implementing new methods for the generics
\code{model.frame} and \code{model.matrix} from the \code{stats} package, and
ensuring that these generics are used by already existing regression fitting
functions.

The main function, \code{\link{nauf_reg}}, is a wrapper function that assigns
the formula for a regression an extra (first) class attribute of
\code{\link[=nauf-class]{nauf}}, and then calls an appropriate regression
fitting function from the \code{stats} or \code{MASS} package, causing these
functions to use the \code{nauf} generics for \code{model.frame} and
\code{model.matrix} (see \code{\link{nauf_model_frame}} and
\code{\link{nauf_model_matrix}}).  Currently, any regression which can be
fit with the functions \code{\link[stats]{lm}}, \code{\link[stats]{glm}},
and \code{\link[MASS]{glm.nb}} can be fit using \code{\link{nauf_reg}} to
allow \code{NA} values in the unordered factors.  For the treatment of
unordered factors, see \code{\link{named_contr_sum}},
\code{\link{nauf_interaction}}, and \code{\link{nauf_contrasts}}.
}

