% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.nauf.R
\name{is.nauf}
\alias{is.nauf}
\alias{is.nauf.formula}
\alias{is.nauf.frame}
\alias{is.nauf.glm}
\alias{is.nauf.glmerMod}
\alias{is.nauf.lmerMod}
\alias{is.nauf.merMod}
\alias{is.nauf.model}
\alias{is.nauf.nested.anova}
\alias{is.nauf.ref.grid}
\alias{is.nauf.terms}
\title{Determine if an object inherits from a \code{nauf} class.}
\usage{
is.nauf.formula(object)

is.nauf.terms(object)

is.nauf.frame(object)

is.nauf.glm(object)

is.nauf.lmerMod(object)

is.nauf.glmerMod(object)

is.nauf.merMod(object)

is.nauf.model(object)

is.nauf.ref.grid(object)

is.nauf.nested.anova(object)
}
\arguments{
\item{object}{Any R object.}
}
\description{
The \code{is.nauf} functions return \code{TRUE} if \code{object} inherits
from the class (or one of the classes) implied in the name of the funciton
and \code{FALSE} otherwise.
}
\details{
The following inheritances will result in a value of \code{TRUE}:
\describe{
  \item{is.nauf.formula}{\code{\linkS4class{nauf.formula}}.}
  \item{is.nauf.terms}{\code{\linkS4class{nauf.terms}}.}
  \item{is.nauf.frame}{\code{\linkS4class{nauf.frame}}.}
  \item{is.nauf.glm}{\code{\linkS4class{nauf.glm}}.}
  \item{is.nauf.lmerMod}{\code{\linkS4class{nauf.lmerMod}}.}
  \item{is.nauf.glmerMod}{\code{\linkS4class{nauf.glmerMod}}.}
  \item{is.nauf.merMod}{\code{\linkS4class{nauf.lmerMod}} or
    \code{\linkS4class{nauf.glmerMod}}.}
  \item{is.nauf.model}{\code{\linkS4class{nauf.glm}},
    \code{\linkS4class{nauf.lmerMod}}, or
    \code{\linkS4class{nauf.glmerMod}}.}
  \item{is.nauf.ref.grid}{\code{\linkS4class{nauf.ref.grid}}.}
  \item{is.nauf.nested.anova}{\code{\linkS4class{nauf.nested.anova}}.}
}
}

