% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_functions.R
\name{named_contr_sum}
\alias{named_contr_sum}
\title{Create named sum contrasts for a factor.}
\usage{
named_contr_sum(x, return_contr = TRUE)
}
\arguments{
\item{x}{A factor or vector, with at least two unique non-\code{NA} values,
which can be coerced to character with \code{\link[base]{as.character}}.}

\item{return_contr}{A logical. If \code{TRUE} (the default), a contrast
matrix is returned. If \code{FALSE}, \code{x} is converted to an unordered
factor with the contrast matrix applied, and the factor is returned.}
}
\value{
If \code{return_contr = TRUE}, a contrast matrix obtained from
  \code{\link[stats]{contr.sum}} with named columns rather than numbered
  columns.  If \code{return_contr = FALSE}, then \code{x} is returned
  as an unordered factor with the named sum contrasts applied.
}
\description{
\code{named_contr_sum} creates sum contrasts for a factor which are named
with the levels of the factor rather than with number (e.g. if a factor
\code{f1} has levels \code{A}, \code{B}, and \code{C}, then rather than
creating contrast columns \code{f11} and \code{f12}, it creates columns
\code{f1A} and \code{f1B}).
}
\details{
First, \code{x} is coerced to character, and its unique non-\code{NA} values
are sorted alphabetically.  If there are two unique values, and they are
equal to (ignoring case) "F" and "T", "FALSE" and "TRUE", "N" and "Y",
"NO" and "YES", or "0" and "1", then their order is reversed (this makes it
so the positive level gets the dummy coefficient rather than the negative
level, which has a more intuitive interpretation).  Then
\code{\link[stats]{contr.sum}} is called, and the column names of the
resulting contrast matrix are set using the character vector of unique values
(excluding the final element that gets coded as \code{-1} for all dummy
variables).  If \code{return_contr = TRUE}, then this contrast matrix is
returned.  If \code{return_contr = FALSE}, then \code{x} is converted to an
unordered factor with the named sum contrats and returned. \code{NA} is never
assigned as a level in the contrast matrix or in the factor returned by the
function, but \code{NA} values in \code{x} are not removed in the factor
returned when \code{return_contr = FALSE}. See the examples.

Using sum contrasts is important for handling \code{NA} values
in a factor since in the model matrix the \code{NAs} can be set to zero for
all of the dummy variables, averaging over the effect of the factor rather
than causing the \code{NAs} to be coded as the base level as would occur
with treatment contrasts (in sociolinguistics this is often referred to as
factor).  In the \code{nauf} package, all unordered factors
are automatically coded with named sum contrasts (as are any variables not
coded as unordered factors, but which have only two unique values; e.g.
logicals, variables coded as integer \code{0 / 1}, etc.).  See
\code{\link{nauf_model_frame}} and \code{\link{nauf_interaction}} for details.
}
\examples{
f <- factor(rep(c("a", "b", "c", NA), 2), levels = c("b", "c", "a"))
f <- addNA(f)
levels(f)  # NA listed as factor
contrasts(f)  # NA included in contrast matrix
named_contr_sum(f)  # named sum contrasts (NA dropped; levels alphabetized)
named_contr_sum(levels(f))  # same output
named_contr_sum(f, FALSE)  # f (values unchanged) with named sum contrasts

f <- c(TRUE, FALSE, FALSE, TRUE)
class(f)  # logical
named_contr_sum(f)  # TRUE gets the dummy variable
f <- named_contr_sum(f, FALSE)
class(f)  # factor

named_contr_sum(letters[1:5])  # character argument
named_contr_sum(rep(letters[1:5], 2), FALSE)  # creates factor

named_contr_sum(c(1, 6, 3))  # numeric argument
named_contr_sum(rep(c(1, 6, 3), 2), FALSE)  # creates factor

# ordered factors are converted to unordered factors, so use with caution
f <- factor(rep(1:3, 2), ordered = TRUE)
is.ordered(f)
f
f <- named_contr_sum(f, FALSE)
is.ordered(f)
f

\dontrun{
# error from stats::contr.sum because only one unique non-NA value
named_contr_sum(5)
named_contr_sum(rep(c("a", NA), 3), FALSE)
}

}
\seealso{
\code{\link{nauf_interaction}} for interactions involving unordered
  factors where at least one has \code{NA} values,
  \code{\link{nauf_model_frame}} for automatic assignment of unordered factor
  contrasts, and \code{\link{nauf_model_matrix}} for the treatment of
  \code{NAs} in regressions.
}

