% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf-class.R
\name{nauf-class}
\alias{is.nauf}
\alias{nauf-class}
\alias{nauf_off}
\alias{nauf_on}
\title{S3 class for fitting models with non-applicable unordered factor levels.}
\usage{
is.nauf(object)

nauf_on(object)

nauf_off(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
The \code{nauf} class is used to ensure that the proper S3 methods
for \code{\link[stats]{model.frame}} and \code{\link[stats]{model.matrix}}
are called by already-existing model fitting functions.
}
\details{
The \code{nauf} package mostly relies on two functions, both of which are
S3 generics: \code{model.frame.nauf} and \code{model.matrix.nauf} with
corresponding wrapper functions \code{\link{nauf_model_frame}} and
\code{\link{nauf_model_matrix}}.

The \code{nauf_on} function adds \code{nauf} as the first
\code{\link[base]{class}} attribute to \code{object} if \code{object}
\code{\link[base]{inherits}} from \code{\link[stats]{formula}},
\code{\link[base]{data.frame}}, \code{list}, \code{\link[stats]{lm}}, or
\code{call} and is not an S4 object.  If the object does not meet these
critera, it is returned unaltered and a warning is issued.

The \code{nauf_off} function removes \code{nauf} from the the class
of \code{object} if it is there and then returns the object.

The \code{is.nauf} function returns \code{TRUE} if \code{object} inherits
from \code{nauf} and \code{FALSE} if it does not.
}

