% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-class.R
\name{nauf.formula-class}
\alias{nauf.formula}
\alias{nauf.formula-class}
\title{Class for \code{nauf} regression formulas.}
\description{
The \code{formula} attribute of a \code{\linkS4class{nauf.frame}} has the
(S3) class \code{nauf.formula} to ensure that calls to 
\code{\link[stats]{model.frame}} which use the formula call 
\code{\link{nauf_model.frame}}.  There are currently no generic methods for 
\code{nauf.formula} objects meant for end user use.
}
\seealso{
\code{\link{nauf_contrasts}}, \code{\link{nauf_model.frame}}, and 
  \code{\link{is.nauf.formula}}.
}

