% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-class.R
\name{nauf.glm-class}
\alias{nauf.glm}
\alias{nauf.glm-class}
\title{Class of fitted fixed effects models with \code{nauf} contrasts.}
\description{
A fitted fixed effects regression model returned by \code{\link{nauf_lm}}, 
\code{\link{nauf_glm}}, or \code{\link{nauf_glm.nb}} with an additional first 
(S3) class attribute \code{nauf.glm}.  There are currently no generic methods 
for class \code{nauf.glm}, and generics from the \code{base}, \code{stats}, 
and \code{MASS} packages such as \code{summary}, \code{predict}, etc. work 
just as they would for models fit with \code{\link[stats]{lm}}, 
\code{\link[stats]{glm}}, or \code{\link[MASS]{glm.nb}}.  If you encounter a 
generic function in these packages which does not function properly, please 
report the issue at \url{https://github.com/CDEager/nauf/issues}.
}
\seealso{
\code{\link{nauf_lm}}, \code{\link{nauf_glm}}, 
  \code{\link{nauf_glm.nb}}, and \code{\link{nauf_contrasts}}
}

