% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-class.R
\docType{class}
\name{nauf.merMod}
\alias{nauf.glmerMod}
\alias{nauf.glmerMod-class}
\alias{nauf.lmerMod}
\alias{nauf.lmerMod-class}
\alias{nauf.merMod}
\title{Class for fitted mixed effects models with \code{nauf} contrasts.}
\description{
Models fit with \code{\link{nauf_lmer}} have class \code{nauf.lmerMod}
(inheriting from \code{\linkS4class{lmerMod}}) and models fit with
\code{\link{nauf_glmer}} and \code{\link{nauf_glmer.nb}} have class
\code{nauf.glmerMod} (inheriting from \code{\linkS4class{glmerMod}}).
}
\section{Slots}{

\describe{
\item{\code{rsp,Gp,call,frame,flist,cnms,lower,theta,beta,u,devcomp,pp,optinfo}}{See
\code{\linkS4class{merMod}}.}
}}
\section{Generic Methods}{

There are S3 methods specific to the
\code{nauf.lmerMod} and \code{nauf.glmerMod} classes for the generic
functions \code{\link[=predict.nauf.merMod]{predict}} and
\code{\link[=anova.nauf.merMod]{anova}}, which behave differently from the
methods for \code{\linkS4class{merMod}} objects. Other methods for generic
functions from the \code{stats}, \code{MASS}, and \code{lme4} packages
should work as they would for \code{\linkS4class{merMod}} objects, with the
exception of the method for the \code{\link[stats]{simulate}} function,
for which the \code{nauf} method has more limited options than
\code{\link[lme4]{simulate.merMod}}, and is not meant for end user use
at this time (the method is necessary for the \code{PB} method to work
for \code{\link[=anova.nauf.merMod]{anova}}). If you encounter a generic
function in these packages which does not function properly, please report
the issue at \url{https://github.com/CDEager/nauf/issues}.
}
\seealso{
\code{\link{nauf_glmer}}, \code{\link{nauf_contrasts}}, and
  \code{\linkS4class{merMod}}.
}

