% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-class.R
\name{nauf.nested.anova-class}
\alias{nauf.nested.anova}
\alias{nauf.nested.anova-class}
\title{Class for \code{nauf} model anovas with nested models.}
\description{
A list returned by \code{\link{anova.nauf.merMod}} when \code{method} is one
of \code{nested-KR}, \code{LRT}, or \code{PB}. Similar to the object returned
by \code{\link[afex]{mixed}}, it has three elements: \code{anova_table} is
an \code{\link[stats]{anova}} table with the significance tests comparing
the full model to each of the nested models; \code{full_model} is the model
with all the fixed effects; and \code{restricted_models} is a named list
of nested models, each lacking one of the fixed effects.
}
\seealso{
\code{\linkS4class{nauf.lmerMod}}, 
  \code{\linkS4class{nauf.glmerMod}}, and \code{\link{anova.nauf.merMod}}.
}

