% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-class.R
\name{nauf.ref.grid-class}
\alias{nauf.ref.grid}
\alias{nauf.ref.grid-class}
\title{Class for reference grids for nauf models}
\description{
A list returned by \code{\link{nauf_ref.grid}} with one element
\code{ref.grid} which is a  \code{\linkS4class{ref.grid}} object. The 
reference grid the \code{nauf.ref.grid} contains should not be manipulated 
directly, or used as an argument to \code{\link[lsmeans]{lsmeans}}.  It 
should only be used with \code{\link{nauf_pmmeans}}.
}

