% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf.stanreg-class-generic.R
\name{nauf.stanreg}
\alias{nauf.stanreg}
\title{Class for fitted Bayesian models with \code{nauf} contrasts.}
\description{
ADD DESCRIPTION
}
\seealso{
\code{\link{nauf_stan_glm}}, \code{\link{nauf_stan_glmer}},
  \code{\link{nauf_contrasts}}, and \code{\link[rstanarm]{stanreg-objects}}.
}

