% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_functions.R
\name{nauf_contrasts}
\alias{nauf_contrasts}
\title{Contrasts for a model fit with \code{\link{nauf_reg}}.}
\usage{
nauf_contrasts(object)
}
\arguments{
\item{object}{A model fit using \code{\link{nauf_reg}}, a model frame made
by \code{\link{nauf_model_frame}}, or the \code{terms} object from either
of these.}
}
\value{
A list with two elements: \code{mefc} (main effect factor contrasts)
  and \code{ccna} (contrast changes due to \code{NAs}).

  \code{mefc} is a list with two elements \code{unordered} and \code{ordered},
  each of which is a named list with an entry for each unordered or ordered
  factor in the regression which contains the contrasts applied to the main
  effect.

  \code{ccna} is a named list with an element for each combination of
  unordered factors which required a change in factor contrasts from the
  main effects, with an entry for each of the factors containing the
  contrasts applied to any term involving those unordered factors (and
  possibly also ordered factors and covariates), but no other unordered
  factors.
}
\description{
\code{nauf_contrasts} lists all factor contrasts in a model fit with
\code{\link{nauf_reg}}, including changes in contrasts for interactions
involving \code{NAs}.
}
\examples{
\dontrun{
m <- nauf_reg(y ~ f1 * f2 * (f3 + x1 + x2), data = mydata)
nauf_contrasts(m)
}

}

