% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf_pmmeans.R
\name{nauf_grid}
\alias{nauf_grid}
\title{Create a reference grid with \code{NA} unordered factors.}
\usage{
nauf_grid(mod)
}
\arguments{
\item{mod}{A regression model fit with \code{\link{nauf_reg}}.}
}
\value{
A \code{\link[lsmeans]{ref.grid-class}} object which includes
  \code{NA} values for relevant unordered factors in the grid expansion.
}
\description{
\code{nauf_grid} creates a reference grid for a model fit with
\code{\link{nauf_reg}} which can be used to obtain predicted marginal
means and pairwise comparisons for factors with \code{\link{nauf_pmmeans}}.
See \code{\link[lsmeans]{ref.grid}} for details of reference grid properties.
}
\details{
A reference grid creates a data frame which contains all possible
combinations of the factors in a regression model, holding all covariates
at their mean values.  There are many options for
\code{\link[lsmeans]{ref.grid}} which are not currently supported for
\code{nauf} models.  The main functionality which is not currently supported
is that the reference grid cannot be created specifying certain levels for
factors (i.e. the \code{at} argument), and the behavior for polynomial
covariates has not been tested.  A direct call to
\code{\link[lsmeans]{ref.grid}} will result in warnings (or possibly errors),
and inference made with the resulting object will be misleading and/or
incorrect.  Only \code{nauf_grid} should be used.

The reference grid returned by \code{nauf_grid} contains combinations of
factors which are not actually possible given the data set.  For example,
if factor \code{f1} has levels \code{A} and \code{B}, and factor \code{f2}
is \code{NA} when \code{f1 = A}, and takes values \code{C} and \code{D}
when \code{f1 = B}, the reference grid will still contain the combinations
\code{f1 = A, f2 = C}, \code{f1 = A, f2 = D}, and \code{f1 = B, f2 = NA},
even though these combinations are not possible.  This is because it is
impossible to know without the researcher's knowledge which combinations
make sense.  In many cases, this is inconsequential for the computation
of predicted marginal means, since the coding of unordered factors in
\code{nauf} regressions will average over the effects.  In cases where
these rows in the reference grid will cause invalid estimates and pairwise
comparisons, subsetting arguments can be used in the call to
\code{\link{nauf_pmmeans}} to ensure only the correct subsets are considered.
}
\examples{
\dontrun{
mod <- nauf_reg(y ~ f1 + f2 + x)
rg <- nauf_grid(mod)
}

}
\seealso{
\code{\link{nauf_pmmeans}}
}

