% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_functions.R
\name{nauf_interaction}
\alias{nauf_interaction}
\title{Unordered factor interactions involving \code{NAs}.}
\usage{
nauf_interaction(x, cols = colnames(x))
}
\arguments{
\item{x}{A data.frame.}

\item{cols}{A vector specifying columns in \code{x} involved in an
interaction.  At least two must be unordered factors.  Defaults to
all columns in \code{x}.}
}
\value{
A list with two elements.  The first, \code{levels}, is a named list
  with one entry for each unordered factor in \code{cols}.  Each entry is a
  character vector of the factor levels which should be coded for in the full
  interaction term (with sum contrasts).  The second element, \code{changed},
  is a logical indicating whether or not \code{levels} is the same as the
  levels in the main effects.
}
\description{
The function determines whether unordered factor levels need to be dropped
in the computation of an interaction term due to \code{NA} values, treating
characters and logicals as unordered factors.
}
\details{
If there are not at least two unordered factors in the specified columns,
an error is thrown. Otherwise, the unordered factors involved in the
interaction are subsetted.  Then a table of all their unique combinations in
\code{x} is made, excluding any combination which contains \code{NAs}.
Then levels which are not used in the no-\code{NA} subset are dropped.
If any of the factors has only one remaining level or if there are any
zeros in the full contingency table of the factors, then a warning is issued
indicating that there is collinearity in the interaction term (the term
shouldn't be included and will cause \code{NA} regression coefficients).
The levels for each factor which are applicable in the interaction are
returned along with a logical indicating if any levels were dropped. Note
that only the highest order interaction of the unordered factors is
considered.  If unordered factors \code{A}, \code{B}, and \code{C} are
involved in an interaction, it may be that \code{A} needs levels dropped in
the third-order interaction, but not in the interaction \code{A:B}, for
example.  The function would need to be called separately with
\code{cols = c("A", "B")} to determine the levels for the second-order
interaction.

This function is implemented in \code{\link{nauf_model_frame}}, and is
necessary in cases where \code{NA} values are collinear with non-\code{NA}
values in other unordered factors.  For example, consider a linguistic
study where participants speak one of three different dialects,
which we will denote \code{A}, \code{B}, and \code{C}.  Suppose that in
dialects \code{A} and \code{B}, some speakers are bilingual in the (same)
second language while others are monolingual, but in dialect \code{C}, all
speakers are monolingual.  In this case, we can code two factors.  The
factor \code{dialect} has levels \code{A}, \code{B}, and \code{C}.  The
factor \code{bilingual} has levels \code{TRUE} and \code{FALSE}, with all
observations pertaining to \code{dialect C} coded as \code{NA}, since the
the factor is not contrastive within the group (i.e. being monolingual in
\code{dialect C} is different from being monolingual in the other two).
Using \code{\link{named_contr_sum}}, we assign the following contrasts (in
the \code{bilingual} table, the \code{NA} row is shown here to be explicit;
it is not a part of the contrast matrix returned by the function because
the \code{\link[stats]{contrasts}} requires the number of rows to be one
more than the number of columns; the \code{NAs} are set to zero by
\code{\link{nauf_model_matrix}}):

Contrasts for main effect of \code{dialect}
\tabular{rr}{
    \tab A \tab B\cr
  A \tab  1 \tab  0\cr
  B \tab  0 \tab  1\cr
  C \tab -1 \tab -1
}

Contrasts for main effect of \code{bilingual}
\tabular{r}{
        \tab TRUE\cr
  TRUE  \tab  1\cr
  FALSE \tab -1\cr
  NA    \tab  0
}

This setup allows the regression coefficient \code{bilingualTRUE} to
only apply when either \code{dialectA} or \code{dialectB} is \code{1},
and to always be multiplied by \code{0} when \code{dialectA} and
\code{dialectB} are \code{-1} (i.e. for \code{dialect C}).  In the regression
output, the predicted grand mean for each dialect is obtained by applying
the contrast coding to the \code{dialect} coefficients.  That is, we have:

\code{
  mean(A) = (Intercept) + dialectA
  mean(B) = (Intercept) + dialectB
  mean(C) = (Intercept) - dialectA - dialectB
}

For dialects \code{A} and \code{B}, this estimate averages over the effect
of \code{bilingual}, and the estimates for the sub-groups can be obtained by:

\code{
  mean(A:TRUE)  = mean(A) + bilingualTRUE
  mean(A:FALSE) = mean(A) - bilingualTRUE
  mean(B:TRUE)  = mean(B) + bilingualTRUE
  mean(B:FALSE) = mean(B) - bilingualTRUE
}

Provided that all covariates have been put on unit scale (see
\code{\link[base]{scale}}) and orthogonal polynomial contrasts have been set
for all ordered factors (see \code{\link[stats]{contr.poly}}), both of which
are good ideas for most regressions anyway, then these types of contrasts
also result in the intercept being the corrected mean, and the output is
easier to interpret.  An additional (and more important) advantage to
implementing \code{NA} values, however, comes when factors such as the ones
in this example interact.  Assume now that we are interested in the
interaction \code{dialect * bilingual}, rather than just the main effects.
Coding the observations from dialect \code{C} as \code{bilingual = FALSE}
yields a rank-deficient matrix (i.e. there are collinear columns).  If we
simply multiply the contrasts for the main effects outlined above in the
interaction (which is what the default for \code{link[stats]{model.matrix}}
does), the result is not collinear; hoever, it is still undesirable:

Undesirable contrasts for \code{dialect:bilingual} interaction
\tabular{llrr}{
   \tab  \tab dialectA:bilingualTRUE \tab dialectB:bilingualTRUE\cr
  A \tab TRUE  \tab  1 \tab  0\cr
  A \tab FALSE \tab -1 \tab  0\cr
  B \tab TRUE  \tab  0 \tab  1\cr
  B \tab FALSE \tab  0 \tab -1\cr
  C \tab NA    \tab  0 \tab  0
}

These contrasts are undesirable because the same interaction term could be
expressed with one column with some releveling, and in cases more complicated
than this one, there could also be collinearity.  So, instead, we should use
the following contrasts:

\tabular{llr}{
   \tab  \tab dialectA:bilingualTRUE\cr
  A \tab TRUE  \tab  1\cr
  A \tab FALSE \tab -1\cr
  B \tab TRUE  \tab -1\cr
  B \tab FALSE \tab  1\cr
  C \tab NA    \tab  0
}

Determining these contrasts manually can be tedious (especially as the number
of factors and levels grows), hence \code{nauf_interaction}.
For this example \code{nauf_interaction} determines that \code{C} is
redundant in the interaction term, and so it drops it from the levels
of \code{dialect}, while keeping \code{bilingual} the same.  In this case,
\code{nauf_interaction} would return a list with an element \code{levels}
which has sub-elements \code{dialect} and \code{bilingual}, each of which is
a character vector (\code{c("A", "B")} and \code{c("TRUE", "FALSE")},
respectively), along with a second element \code{changed = TRUE} to indicate
that at least one level was dropped from at least one factor.  See
\code{\link{nauf_model_matrix}} for implementation of interaction contrasts
which don't match main effect contrasts, and how this effects interpretation
of the regression output.  See the examples section for code
implementing this example.
}
\examples{
dat <- data.frame(
  dialect = c("A", "A", "B", "B", "C"),
  bilingual = c(TRUE, FALSE, TRUE, FALSE, NA))
nauf_interaction(dat)  # drops dialect C

dat <- as.data.frame(lapply(dat, function(n) rep(n, 10)))
dat$x <- rnorm(nrow(dat))
dat$o <- factor(rep(1:3, length.out = 50), ordered = TRUE)
nauf_interaction(dat)  # ignores x and o; same result as above

\dontrun{
nauf_interaction(dat, "dialect")  # error; only one column
nauf_interaction(dat, c("dialect", "x"))  # error; only one unordered factor
}

}
\seealso{
\code{\link{named_sum_contr}} for the contrasts that \code{nauf}
  assigns to unordered factors,
  \code{\link{nauf_model_frame}} for automatic assignment of unordered factor
  contrasts, and \code{\link{nauf_model_matrix}} for the treatment of
  \code{NAs} in regressions.
}

