% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf.stanreg-class-generic.R
\name{nauf_kfold}
\alias{nauf_kfold}
\title{Cross-validation for \code{nauf.stanreg} models.}
\usage{
nauf_kfold(x, K = 10, save_fits = FALSE)
}
\arguments{
\item{x, K, save_fits}{See \code{\link[rstanarm]{kfold}}.}
}
\value{
An object with classes \code{kfold} and \code{loo}.
}
\description{
The same as \code{\link[rstanarm]{kfold}}, but ensuring the use of
\code{\link{nauf_contrasts}}.
}

