% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf.stanreg-class-generic.R
\name{nauf_launch_shinystan}
\alias{nauf_launch_shinystan}
\title{Launch the ShinyStan app for a \code{nauf} model.}
\usage{
nauf_launch_shinystan(object, rstudio = getOption("shinystan.rstudio"), ...)
}
\arguments{
\item{object}{An object of class \code{\link{nauf.stanreg}}.}

\item{rstudio, ...}{See \code{\link[shinystan]{launch_shinystan}}.}
}
\value{
See \code{\link[shinystan]{launch_shinystan}}.
}
\description{
\code{nauf_launch_shinystan} is just a wrapper function for
\code{\link[shinystan]{launch_shinystan}} which removes the \code{nauf.stanreg}
class attribute from the \code{object} argument.
}

