% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf_model_frame.R
\name{nauf_model_frame}
\alias{model.frame.nauf}
\alias{nauf_model_frame}
\title{Get the model frame for a model with unordered factor \code{NA} values.}
\usage{
nauf_model_frame(formula, data = NULL, na.action = "na.pass",
  drop.unused.levels = TRUE, xlev = NULL, ...)

\method{model.frame}{nauf}(formula, data = NULL, na.action = "na.pass",
  drop.unused.levels = TRUE, xlev = NULL, ...)
}
\arguments{
\item{formula}{A \code{\link[stats]{formula}}. \code{\link[stats]{terms}}
attributes are ignored.}

\item{data}{A data.frame containing the variables for the model.}

\item{na.action}{Changes from the default value (\code{\link[stats]{na.pass}})
are ignored (\code{NA} values in factors are not removed).}

\item{drop.unused.levels}{Changes from the default value (\code{FALSE}) are
ignored (factor levels with no observations are dropped).}

\item{xlev}{Changes from the default value (\code{NULL}) are ignored. Levels
for unordered factors are determined alphabetically so they are consistent
when reordered in interactions (see \code{\link{nauf_interaction}}).}

\item{...}{Other parameters to pass to the default
\code{\link[stats]{model.frame}} method such as \code{subset},
\code{weights} and \code{offset}.}
}
\value{
A model frame that inherits from \code{\link[=nauf-class]{nauf}} and
  whose \code{\link[stats]{terms}} attribute inherits from
  \code{\link[=nauf-class]{nauf}} and has three additional attributes:
  \code{mefc} (main effect factor contrasts) is a named list with an element
  for each factor in the model, containing its contrast matrix; \code{ccna}
  (contrasts changed due to \code{NA} values) is a named list with an entry
  for each unordered factor interaction which requires contrasts to be
  changed from those applied to the main effects, containing a named list
  of levels which are kept in the interaction (see
  \code{\link{nauf_interaction}}); and \code{hasna} is a named logical vector
  indicating which variables in the model frame have \code{NA} values.
}
\description{
\code{nauf_model_frame} calls the default \code{\link[stats]{model.frame}}
method with \code{na.action = "na.pass"} and then sets sum contrasts for
unordered factors and checks for \code{NAs} that cause factor levels to
need to be dropped in interactions.
}
\details{
After calling the default \code{\link[stats]{model.frame}} method,
the function converts any character and logical columns to unordered factors,
and any other column which has only two unique non-\code{NA} values,
regardless of their class, are also coerced to unordered factors (e.g.
a factor coded as ordered even though it only has two levels, or a binary
variable coded as integer \code{0 / 1}).  Then all unordered factors are
assigned named sum contrasts via \code{\link{named_contr_sum}}.

After this, all columns are checked for \code{NA} values.  If any columns
have \code{NAs} but are not unordered factors, then an error is thrown
(these types of \code{NA} values are not the same as the ones dealt with in
this package; there are many options for imputing them, but this needs to be
done prior to calling any \code{nauf} functions).  If there are any unordered
factors with \code{NA} values which are involved in interactions with other
unordered factors, then for each of these interactions
\code{\link{nauf_interaction}} is called, and, if changes in contrasts are
required in the interaction term, they are recorded.
}
\section{Methods (by generic)}{
\itemize{
\item \code{model.frame}: S3 method for class 'nauf'
}}
\examples{
\dontrun{
fr <- nauf_model_frame(y ~ f1 * f2 * (f3 + x1), mydata)
}

}

