% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf_model_matrix.R
\name{nauf_model_matrix}
\alias{model.matrix.nauf}
\alias{nauf_model_matrix}
\title{Get the model matrix for a model with unordered factor \code{NA} values.}
\usage{
nauf_model_matrix(object = NULL, data = NULL, ...)

\method{model.matrix}{nauf}(object = NULL, data = NULL, ...)
}
\arguments{
\item{object}{A model frame fit by \code{\link{nauf_model_frame}}, a
regression model fit by \code{\link{nauf_reg}}, a
\code{link[stats]{formula}}, or \code{NULL}.  See 'Details'.}

\item{data}{A data.frame or \code{NULL}.  See 'Details'.}

\item{...}{Additional arguments to pass to \code{\link{nauf_model_frame}}}
}
\value{
A model matrix with unordered factor \code{NAs} set to zero.  It
  has an \code{assign} attribute, but no \code{contrasts} attribute.
}
\description{
\code{nauf_model_matrix} extracts the model matrix from a regression fit
with \code{\link{nauf_reg}}, constructs the model matrix specified in
a model frame built with \code{\link{nauf_model_frame}}, or constructs
a model matrix after calling \code{\link{nauf_model_frame}}, depending
on the value of \code{object} and \code{data}.  See 'Details'.
}
\details{
Exactly what happens depends on the value of the arguments \code{object}
and \code{data}.  The following are organized in order of evaluation:

If \code{object} inherits from \code{\link[=nauf-class]{nauf}} and \code{lm}
(i.e. it is a regression fit with \code{\link{nauf_reg}}), then the model
matrix stored in the regression fit is returned (it is the same as simply
calling \code{object$x}).

If \code{object} inherits from \code{\link[=nauf-class]{nauf}} and
\code{data.frame} (i.e. it is a model frame created by
\code{\link{nauf_model_frame}}), then the function begins by calling the
default method for \code{\link[stats]{model.matrix}}, and then checking
whether the \code{ccna} attribute of the model frame's terms object indicates
that any unordered factor interactions require a change in contrasts (see
\code{\link{nauf_interaction}}).  For each interaction which requires changes,
a copy of the model frame is made and in the copy the factor is re-leveled,
dropping levels which are not applicable in the interaction, and assigning
contrasts to the new factor with \code{\link{named_contr_sum}}.  Then
a model matrix for all terms which contain the relevant unordered factor
interaction is generated, and the columns in the original model matrix
are replaced with the revised contrasts, appending "ccna_" at the beginning
of the column names so that they are easily spotted in the regression output.
Finally, all \code{NAs} in the model matrix are set to zero and the matrix
is returned.

If \code{object} does not meet either of the above specifications, then
\code{data} is checked.  If it is a model frame created by
\code{\link{nauf_model_frame}}, then a model matrix is created in the way
just described.

If none of the previous three conditions are met, and \code{object} is a
\code{\link[stats]{formula}} and \code{data} is a \code{data.frame}, then
\code{\link{nauf_model_frame}} is called, passing along any additional
arguments in \code{...}, and a model matrix is built with the resulting
model frame.
}
\section{Methods (by generic)}{
\itemize{
\item \code{model.matrix}: S3 method for class 'nauf'
}}
\examples{
\dontrun{
fr <- nauf_model_frame(y ~ f1 * f2 * (f3 + x1), mydata)
mod <- nauf_reg(y ~ f1 * f2 * (f3 + x1), mydata)

# all of the following return the same
x <- nauf_model_matrix(y ~ f1 * f2 * (f3 + x1), mydata)
x <- nauf_model_matrix(fr)
x <- nauf_model_matrix(data = fr)
x <- nauf_model_matrix(mod)
}

}

