% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf_pmmeans.R
\name{nauf_pmmeans}
\alias{nauf_pmmeans}
\title{Predicted marginal means for nauf models.}
\usage{
nauf_pmmeans(object, specs, keep_level = list(), drop_level = list(),
  keep_group = list(), drop_group = list(), ...)
}
\arguments{
\item{object}{A \code{\linkS4class{nauf.ref.grid}} or any \code{nauf} model
(see \code{\link{is.nauf.model}}).}
}
\description{
ADD DESCRIPTION
}
\details{
ADD DETAILS
}

