% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf_pmmeans.R
\name{nauf_ref.grid}
\alias{nauf_ref.grid}
\title{Create a reference grid for a nauf model.}
\usage{
nauf_ref.grid(mod)
}
\arguments{
\item{mod}{Any \code{nauf} model (see \code{\link{is.nauf.model}}).}
}
\description{
Serves the same purpose as \code{\link[lsmeans]{ref.grid}}, but only takes
one argument (a \code{nauf} model).  See 'Details'.
}
\details{
ADD DETAILS
}

