% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf_stan_glmer.R
\name{nauf_stan_glmer}
\alias{nauf_stan_glmer}
\alias{nauf_stan_glmer.nb}
\alias{nauf_stan_lmer}
\title{Fit a Bayesian mixed effects regression with \code{nauf} contrasts.}
\usage{
nauf_stan_glmer(formula, data = NULL, family = gaussian, subset, weights,
  na.action = na.pass, offset, contrasts = NULL, ncs_scale = attr(formula,
  "standardized.scale"), ..., prior = rstanarm::normal(),
  prior_intercept = rstanarm::normal(), prior_aux = rstanarm::cauchy(0, 5),
  prior_covariance = rstanarm::decov(), prior_PD = FALSE,
  algorithm = "sampling", adapt_delta = NULL, QR = FALSE,
  sparse = FALSE)

nauf_stan_lmer(formula, data = NULL, subset, weights, na.action = na.pass,
  offset, contrasts = NULL, ncs_scale = attr(formula, "standardized.scale"),
  ..., prior = rstanarm::normal(), prior_intercept = rstanarm::normal(),
  prior_aux = rstanarm::cauchy(0, 5), prior_covariance = rstanarm::decov(),
  prior_PD = FALSE, algorithm = "sampling", adapt_delta = NULL,
  QR = FALSE)

nauf_stan_glmer.nb(formula, data = NULL, subset, weights,
  na.action = na.pass, offset, contrasts = NULL, link = "log",
  ncs_scale = attr(formula, "standardized.scale"), ...,
  prior = rstanarm::normal(), prior_intercept = rstanarm::normal(),
  prior_aux = rstanarm::cauchy(0, 5), prior_covariance = rstanarm::decov(),
  prior_PD = FALSE, algorithm = "sampling", adapt_delta = NULL,
  QR = FALSE)
}
\arguments{
\item{formula, data, family, subset, weights, na.action, offset, contrasts, ncs_scale}{See
\code{\link{nauf_model.frame}} and \code{\link{nauf_glFormula}}.}

\item{...}{Further arguments to be passed to \code{\link[rstan]{sampling}}.
See \code{\link[rstanarm]{stan_glmer}} for details.}

\item{algorithm}{Changes from the default \code{"sampling"} result in an
error.  Only MCMC is currently supported.}

\item{link, prior, prior_intercept, prior_aux, prior_covariance, prior_PD, adapt_delta, QR, sparse}{See
\code{\link[rstanarm]{stan_glmer}}.}
}
\value{
A \code{\link{nauf.stanreg}} object.
}
\description{
ADD DESCRIPTION
}
\details{
ADD DETAILS
}
\examples{
\dontrun{
dat <- fricatives
# add NA stuff
mod <- nauf_stan_lmer(f, d)
}

}
\seealso{
\code{\link{nauf_contrasts}} for a description of the treatment of 
  \code{NA} values, \code{\link[rstanarm]{stan_glmer}} for a description of 
  the priors, and the documentation for \code{Stan} and the \code{rstan} and 
  \code{rstanarm} packages for algorithmic details.
}

