% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-class.R
\docType{class}
\name{nauf.lmerMod-class}
\alias{nauf.lmerMod}
\alias{nauf.lmerMod-class}
\title{Class for fitted linear mixed effects models with \code{nauf} contrasts.}
\description{
A fitted mixed effects regression model returned by \code{\link{nauf_lmer}}
of S4 class \code{nauf.lmerMod}, which inherits from
\code{\linkS4class{lmerMod}}.
}
\details{
There are S3 methods for
the \code{\link[=predict.nauf.merMod]{predict}} and 
\code{\link[=anova.nauf.merMod]{anova}} functions meant for end user use.
There is also an S3 method for \code{\link[stats]{simulate}}, but with more
limited functionality than \code{\link[lme4]{simulate.merMod}}; the
\code{nauf.lmerMod} method for this function is not meant for end users, and
is necessary for the \code{PB} method to work for
\code{\link[=anova.nauf.merMod]{anova}}.
Other generics from the \code{base}, \code{stats}, and
\code{lme4} packages such as \code{summary} etc. work just as
they would for models fit with \code{\link[lme4]{lmer}}.  If you encounter a
generic function in these packages which does not function properly, please
report the issue at \url{https://github.com/CDEager/nauf/issues}.
}
\section{Slots}{

\describe{
\item{\code{rsp,Gp,call,frame,flist,cnms,lower,theta,beta,u,devcomp,pp,optinfo}}{See
\code{\linkS4class{merMod}}.}
}}
\seealso{
\code{\link{nauf_lmer}} and \code{\link{nauf_contrasts}}.
}

