% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf_pmmeans.R
\name{nauf_ref.grid}
\alias{nauf_ref.grid}
\title{Create a reference grid for a \code{nauf} model.}
\usage{
nauf_ref.grid(mod, KR = NULL)
}
\arguments{
\item{mod}{A regression model fit with \code{\link{nauf_reg}}.}

\item{KR}{Only applies when \code{mod} is a \code{\linkS4class{nauf.lmerMod}}
fit with \code{REML = TRUE}. If \code{KR = TRUE}, then the Kenward-Roger
approximation is used to calculate degrees of freedom.  If 
\code{KR = FALSE}, then the Satterthwaite approximation is used.  If
\code{KR = NULL} (the default), then the Kenward-Roger approximation is
used when there are \code{3000} observations or fewer in the model, and the
Satterthwaite approximation is used when there are more than \code{3000}
observations (since in this case the the Kenward-Roger approximation
requires a lot of computation time).  When \code{mod} is a
\code{\linkS4class{nauf.lmerMod}} fit with \code{REML = FALSE}, then the
Satterthwaite approximation is always used.  The Kenward-Roger method
is implemented with \code{\link[pbkrtest]{Lb_ddf}} and the Satterthwaite
method is implemented with \code{\link[lmerTest]{calcSatterth}}.}
}
\value{
A \code{\linkS4class{nauf.ref.grid}}.
}
\description{
\code{nauf_ref.grid} creates a reference grid for a model fit with
\code{\link{nauf_contrasts}} which can be used to obtain predicted marginal
means and pairwise comparisons for factors with \code{\link{nauf_pmmeans}}.
See \code{\link[lsmeans]{ref.grid}} for details of reference grid properties.
}
\details{
A reference grid creates a data frame which contains all possible
combinations of the factors in a regression model, holding all covariates
at their mean values.  There are many options for
\code{\link[lsmeans]{ref.grid}} which are not currently supported for
\code{nauf} models.  The main functionality which is not currently supported
is that the reference grid cannot be created specifying certain levels for
variables (i.e. the \code{at} argument; but see the various subsetting options
in \code{\link{nauf_pmmeans}}).  For models with polynomial covariates,
the results may be misleading, as is usually the case.  A direct call to
\code{\link[lsmeans]{ref.grid}} will result in warnings (or possibly errors),
and inference made with the resulting object will be misleading and/or
incorrect.  Only \code{nauf_ref.grid} should be used.

The reference grid returned by \code{nauf_grid} contains combinations of
factors which are not actually possible in the data set.  For example,
if factor \code{f1} has levels \code{A} and \code{B}, and factor \code{f2}
is \code{NA} when \code{f1 = A}, and takes values \code{C} and \code{D}
when \code{f1 = B}, the reference grid will still contain the combinations
\code{f1 = A, f2 = C}; \code{f1 = A, f2 = D}; and \code{f1 = B, f2 = NA},
even though these combinations are not possible.  This is because it is
impossible to know without the user's knowledge which combinations
make sense.  In many cases, this is inconsequential for the computation
of predicted marginal means, since the coding of unordered factors in
\code{nauf} regressions will average over the effects.  In cases where
these rows in the reference grid will cause invalid estimates and pairwise
comparisons, subsetting arguments can be used in the call to
\code{\link{nauf_pmmeans}} to ensure only the correct subsets are considered.
}
\examples{
\dontrun{
mod <- nauf_lm(y ~ f1 + f2 + x)
rg <- nauf_ref.grid(mod)
}

}
\seealso{
\code{\link{nauf_pmmeans}}
}

