% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nauf_stan_regs.R
\name{nauf_stan_glm}
\alias{nauf_stan_glm}
\alias{nauf_stan_glm.nb}
\alias{nauf_stan_lm}
\title{Fit a Bayesian fixed effects regression with \code{nauf} contrasts.}
\usage{
nauf_stan_glm(formula, family = gaussian(), data = NULL, weights, subset,
  na.action = na.pass, offset = NULL, model = TRUE, x = TRUE,
  y = TRUE, contrasts = NULL, ncs_scale = attr(formula,
  "standardized.scale"), ..., prior = rstanarm::normal(),
  prior_intercept = rstanarm::normal(), prior_aux = rstanarm::cauchy(0, 5),
  prior_PD = FALSE, algorithm = "sampling", adapt_delta = NULL,
  QR = FALSE, sparse = FALSE)

nauf_stan_lm(formula, data = NULL, subset, weights, na.action = na.pass,
  model = TRUE, x = TRUE, y = TRUE, singular.ok = TRUE,
  contrasts = NULL, offset, ncs_scale = attr(formula, "standardized.scale"),
  ..., prior = rstanarm::R2(stop("'location' must be specified")),
  prior_intercept = NULL, prior_PD = FALSE, algorithm = "sampling",
  adapt_delta = NULL)

nauf_stan_glm.nb(formula, data = NULL, weights, subset, na.action = na.pass,
  offset = NULL, model = TRUE, x = TRUE, y = TRUE, contrasts = NULL,
  link = "log", ncs_scale = attr(formula, "standardized.scale"), ...,
  prior = rstanarm::normal(), prior_intercept = rstanarm::normal(),
  prior_aux = rstanarm::cauchy(0, 5), prior_PD = FALSE,
  algorithm = "sampling", adapt_delta = NULL, QR = FALSE)
}
\arguments{
\item{formula, data, family, subset, weights, na.action, offset, contrasts, model, x, y, ncs_scale}{See
\code{\link{nauf_model.frame}}.}

\item{...}{Further arguments to be passed to \code{\link[rstan]{sampling}}.
See \code{\link[rstanarm]{stan_glmer}} for details.}

\item{algorithm}{Changes from the default \code{"sampling"} result in an
error.  Only MCMC is currently supported.}

\item{link, prior, prior_intercept, prior_aux, prior_PD, adapt_delta, QR, sparse}{See
\code{\link[rstanarm]{stan_glm}}.}
}
\value{
A \code{\link{nauf.stanreg}} object.
}
\description{
ADD DESCRIPTION
}
\details{
ADD DETAILS
}
\examples{
\dontrun{
dat <- fricatives
# add NA stuff
mod <- nauf_stan_lm(f, d)
}

}
\seealso{
\code{\link{nauf_contrasts}} for a description of the treatment of 
  \code{NA} values, \code{\link[rstanarm]{stan_glm}} for a description of 
  the priors, and the documentation for \code{Stan} and the \code{rstan} and 
  \code{rstanarm} packages for algorithmic details.
}

